# src/main/java/com/batchsight/demo/dto/AuditTrailDtoDTO.java
/*
 * iBatchPharma API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


package com.batchsight.demo.dto;

import com.fasterxml.jackson.annotation.JsonProperty;
import io.swagger.annotations.ApiModel;
import io.swagger.annotations.ApiModelProperty;

import java.util.Objects;

/**
 * 审计追踪数据实体
 */
@ApiModel(description = "审计追踪数据实体")
@javax.annotation.Generated(value = "org.openapitools.codegen.languages.JavaJerseyServerCodegen", date = "2019-08-23T09:18:38.531+08:00[Asia/Shanghai]")
public class AuditTrailDtoDTO   {
  public static final String JSON_PROPERTY_ID = "id";
  @JsonProperty(JSON_PROPERTY_ID)
  private Integer id;

  public static final String JSON_PROPERTY_UUID = "uuid";
  @JsonProperty(JSON_PROPERTY_UUID)
  private String uuid;

  public static final String JSON_PROPERTY_LOGIN_USER_ID = "login_user_id";
  @JsonProperty(JSON_PROPERTY_LOGIN_USER_ID)
  private Integer loginUserId;

  public static final String JSON_PROPERTY_SID = "sid";
  @JsonProperty(JSON_PROPERTY_SID)
  private Integer sid;

  public static final String JSON_PROPERTY_REASON = "reason";
  @JsonProperty(JSON_PROPERTY_REASON)
  private String reason;

  public static final String JSON_PROPERTY_CLIENT_ID = "client_id";
  @JsonProperty(JSON_PROPERTY_CLIENT_ID)
  private Integer clientId;

  public static final String JSON_PROPERTY_WORKCENTER_ID = "workcenter_id";
  @JsonProperty(JSON_PROPERTY_WORKCENTER_ID)
  private Integer workcenterId;

  public static final String JSON_PROPERTY_TIMESTAMP = "timestamp";
  @JsonProperty(JSON_PROPERTY_TIMESTAMP)
  private Long timestamp;

  public static final String JSON_PROPERTY_VERSION = "version";
  @JsonProperty(JSON_PROPERTY_VERSION)
  private Integer version;

  public AuditTrailDtoDTO id(Integer id) {
    this.id = id;
    return this;
  }

  /**
   * ID
   * @return id
   **/
  @JsonProperty("id")
  @ApiModelProperty(value = "ID")
  
  public Integer getId() {
    return id;
  }

  public void setId(Integer id) {
    this.id = id;
  }

  public AuditTrailDtoDTO uuid(String uuid) {
    this.uuid = uuid;
    return this;
  }

  /**
   * UUID
   * @return uuid
   **/
  @JsonProperty("uuid")
  @ApiModelProperty(value = "UUID")
  
  public String getUuid() {
    return uuid;
  }

  public void setUuid(String uuid) {
    this.uuid = uuid;
  }

  public AuditTrailDtoDTO loginUserId(Integer loginUserId) {
    this.loginUserId = loginUserId;
    return this;
  }

  /**
   * 登录人ID
   * @return loginUserId
   **/
  @JsonProperty("login_user_id")
  @ApiModelProperty(value = "登录人ID")
  
  public Integer getLoginUserId() {
    return loginUserId;
  }

  public void setLoginUserId(Integer loginUserId) {
    this.loginUserId = loginUserId;
  }

  public AuditTrailDtoDTO sid(Integer sid) {
    this.sid = sid;
    return this;
  }

  /**
   * 签名ID
   * @return sid
   **/
  @JsonProperty("sid")
  @ApiModelProperty(value = "签名ID")
  
  public Integer getSid() {
    return sid;
  }

  public void setSid(Integer sid) {
    this.sid = sid;
  }

  public AuditTrailDtoDTO reason(String reason) {
    this.reason = reason;
    return this;
  }

  /**
   * 原因
   * @return reason
   **/
  @JsonProperty("reason")
  @ApiModelProperty(value = "原因")
  
  public String getReason() {
    return reason;
  }

  public void setReason(String reason) {
    this.reason = reason;
  }

  public AuditTrailDtoDTO clientId(Integer clientId) {
    this.clientId = clientId;
    return this;
  }

  /**
   * 终端ID
   * @return clientId
   **/
  @JsonProperty("client_id")
  @ApiModelProperty(value = "终端ID")
  
  public Integer getClientId() {
    return clientId;
  }

  public void setClientId(Integer clientId) {
    this.clientId = clientId;
  }

  public AuditTrailDtoDTO workcenterId(Integer workcenterId) {
    this.workcenterId = workcenterId;
    return this;
  }

  /**
   * 工位ID
   * @return workcenterId
   **/
  @JsonProperty("workcenter_id")
  @ApiModelProperty(value = "工位ID")
  
  public Integer getWorkcenterId() {
    return workcenterId;
  }

  public void setWorkcenterId(Integer workcenterId) {
    this.workcenterId = workcenterId;
  }

  public AuditTrailDtoDTO timestamp(Long timestamp) {
    this.timestamp = timestamp;
    return this;
  }

  /**
   * 记录时间
   * @return timestamp
   **/
  @JsonProperty("timestamp")
  @ApiModelProperty(value = "记录时间")
  
  public Long getTimestamp() {
    return timestamp;
  }

  public void setTimestamp(Long timestamp) {
    this.timestamp = timestamp;
  }

  public AuditTrailDtoDTO version(Integer version) {
    this.version = version;
    return this;
  }

  /**
   * 版本:用于扩展，不同版本的数据对应不同的程序逻辑
   * @return version
   **/
  @JsonProperty("version")
  @ApiModelProperty(value = "版本:用于扩展，不同版本的数据对应不同的程序逻辑")
  
  public Integer getVersion() {
    return version;
  }

  public void setVersion(Integer version) {
    this.version = version;
  }


  @Override
  public boolean equals(java.lang.Object o) {
    if (this == o) {
      return true;
    }
    if (o == null || getClass() != o.getClass()) {
      return false;
    }
    AuditTrailDtoDTO auditTrailDto = (AuditTrailDtoDTO) o;
    return Objects.equals(this.id, auditTrailDto.id) &&
        Objects.equals(this.uuid, auditTrailDto.uuid) &&
        Objects.equals(this.loginUserId, auditTrailDto.loginUserId) &&
        Objects.equals(this.sid, auditTrailDto.sid) &&
        Objects.equals(this.reason, auditTrailDto.reason) &&
        Objects.equals(this.clientId, auditTrailDto.clientId) &&
        Objects.equals(this.workcenterId, auditTrailDto.workcenterId) &&
        Objects.equals(this.timestamp, auditTrailDto.timestamp) &&
        Objects.equals(this.version, auditTrailDto.version);
  }

  @Override
  public int hashCode() {
    return Objects.hash(id, uuid, loginUserId, sid, reason, clientId, workcenterId, timestamp, version);
  }


  @Override
  public String toString() {
    StringBuilder sb = new StringBuilder();
    sb.append("class AuditTrailDtoDTO {\n");
    
    sb.append("    id: ").append(toIndentedString(id)).append("\n");
    sb.append("    uuid: ").append(toIndentedString(uuid)).append("\n");
    sb.append("    loginUserId: ").append(toIndentedString(loginUserId)).append("\n");
    sb.append("    sid: ").append(toIndentedString(sid)).append("\n");
    sb.append("    reason: ").append(toIndentedString(reason)).append("\n");
    sb.append("    clientId: ").append(toIndentedString(clientId)).append("\n");
    sb.append("    workcenterId: ").append(toIndentedString(workcenterId)).append("\n");
    sb.append("    timestamp: ").append(toIndentedString(timestamp)).append("\n");
    sb.append("    version: ").append(toIndentedString(version)).append("\n");
    sb.append("}");
    return sb.toString();
  }

  /**
   * Convert the given object to string with each line indented by 4 spaces
   * (except the first line).
   */
  private String toIndentedString(java.lang.Object o) {
    if (o == null) {
      return "null";
    }
    return o.toString().replace("\n", "\n    ");
  }
}